Rename Main -> NotyNavigation and Move Screen.kt to view package
